requirements.txt
# HoneyNet Dependencies

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Machine Learning & AI
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
torch==2.1.2

# Security & Cryptography
cryptography==41.0.7
bcrypt==4.1.2
pyjwt==2.8.0
passlib==1.7.4

# Networking & Communication
websockets==12.0
aiohttp==3.9.1
requests==2.31.0
paramiko==3.4.0

# Monitoring & File System
psutil==5.9.6
watchdog==6.0.0
prometheus-client==0.19.0

# Production & Performance
aiohttp==3.9.1
secrets==3.3.2

# Analytics & Geolocation
aiosqlite==0.19.0
geoip2==4.7.0
user-agents==2.2.0
maxminddb==2.2.0

# External Services Integration
sendgrid==6.10.0
structlog==23.2.0
sentry-sdk==1.38.0

# Desktop GUI (for client app)
tkinter-modern==1.0.0
customtkinter==5.2.0
pillow==10.1.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
watchdog==3.0.0
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
